[encryption] Set CRYPT_ACTIVATE_ALLOW_DISCARDS persistent activation flag
dm-crypt doesn't forward TRIM commands by default. The reasoning is that
it might leak some data about the device content (such as information on
unused sectors or filesystem type). See [0] for more details.

On Droidian, these implications do not outweigh the potential gains we can
have by allowing TRIM passthrough, so let's enable this as a persistent flag.

It's to be noted that, for example, the used filesystem is already well-known
on Droidian devices.

[0] https://asalor.blogspot.com/2011/08/trim-dm-crypt-problems.html

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>